backend/requirements.txt
asgiref==3.9.1
astroid==3.3.10
colorama==0.4.6
dill==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-stubs==5.2.1
django-stubs-ext==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
